Renamed `getAnnotableRanges` -> `getAnnotatableRanges`
