+++ packages/text-annotator/src/utils/index.ts
@@ -1,7 +1,7 @@
 export * from './getClientRectsPonyfill';
 export * from './mergeClientRects';
 export * from './trimRange';
-export * from './getAnnotableRanges';
+export * from './getAnnotatableRanges';
 export * from './getQuoteContext';
 export * from './rangeToSelector';
 export * from './reviveAnnotation';
